tests: Bump memory size for 3 sockets VMI
Guest OS memory usage apparently goes up with socket count, and
with the current version of QEMU we can already see that 64M is
barely enough to run this test successfully: it's enough to run
a bunch of times in a row to encounter a failure. Let's increase
the VMI's memory size to 80M to ensure the test succeeds
consistently.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>